fix: priceUsd in json corresponds to decimal, not int

+++ CryptoApp.Core/CryptoCurrencies/PriceDataPoint.cs
@@ -2,6 +2,6 @@
 
 public class PriceDataPoint
 {
-    public int PriceUsd { get; set; }
+    public decimal PriceUsd { get; set; }
     public DateTime Time { get; set; }
 }
